+++ src/main/java/org/mycompany/model/cloudforms/EmsCluster.java
@@ -35,37 +35,37 @@ public class EmsCluster {
     @JsonProperty("name")
     private String name;
     @JsonProperty("aggregate_physical_cpus")
-    private Integer aggregatePhysicalCpus;
+    private Long aggregatePhysicalCpus;
     @JsonProperty("aggregate_cpu_total_cores")
-    private Integer aggregateCpuTotalCores;
+    private Long aggregateCpuTotalCores;
     @JsonProperty("aggregate_cpu_speed")
-    private Integer aggregateCpuSpeed;
+    private Long aggregateCpuSpeed;
     @JsonProperty("aggregate_memory")
-    private Integer aggregateMemory;
+    private Long aggregateMemory;
     @JsonProperty("effective_cpu")
-    private Integer effectiveCpu;
+    private Long effectiveCpu;
     @JsonProperty("effective_memory")
-    private Integer effectiveMemory;
+    private Long effectiveMemory;
     @JsonProperty("aggregate_vm_cpus")
-    private Integer aggregateVmCpus;
+    private Long aggregateVmCpus;
     @JsonProperty("aggregate_vm_memory")
-    private Integer aggregateVmMemory;
+    private Long aggregateVmMemory;
     @JsonProperty("drs_enabled")
     private Boolean drsEnabled;
     @JsonProperty("drs_automation_level")
     private String drsAutomationLevel;
     @JsonProperty("drs_migration_threshold")
-    private Integer drsMigrationThreshold;
+    private Long drsMigrationThreshold;
     @JsonProperty("ha_enabled")
     private Boolean haEnabled;
     @JsonProperty("ha_admit_control")
     private Boolean haAdmitControl;
     @JsonProperty("ha_max_failures")
-    private Integer haMaxFailures;
+    private Long haMaxFailures;
     @JsonProperty("total_direct_vms")
-    private Integer totalDirectVms;
+    private Long totalDirectVms;
     @JsonProperty("total_direct_miq_templates")
-    private Integer totalDirectMiqTemplates;
+    private Long totalDirectMiqTemplates;
     @JsonProperty("v_cpu_vr_ratio")
     private Double vCpuVrRatio;
     @JsonProperty("v_ram_vr_ratio")
@@ -86,82 +86,82 @@ public class EmsCluster {
     }
 
     @JsonProperty("aggregate_physical_cpus")
-    public Integer getAggregatePhysicalCpus() {
+    public Long getAggregatePhysicalCpus() {
         return aggregatePhysicalCpus;
     }
 
     @JsonProperty("aggregate_physical_cpus")
-    public void setAggregatePhysicalCpus(Integer aggregatePhysicalCpus) {
+    public void setAggregatePhysicalCpus(Long aggregatePhysicalCpus) {
         this.aggregatePhysicalCpus = aggregatePhysicalCpus;
     }
 
     @JsonProperty("aggregate_cpu_total_cores")
-    public Integer getAggregateCpuTotalCores() {
+    public Long getAggregateCpuTotalCores() {
         return aggregateCpuTotalCores;
     }
 
     @JsonProperty("aggregate_cpu_total_cores")
-    public void setAggregateCpuTotalCores(Integer aggregateCpuTotalCores) {
+    public void setAggregateCpuTotalCores(Long aggregateCpuTotalCores) {
         this.aggregateCpuTotalCores = aggregateCpuTotalCores;
     }
 
     @JsonProperty("aggregate_cpu_speed")
-    public Integer getAggregateCpuSpeed() {
+    public Long getAggregateCpuSpeed() {
         return aggregateCpuSpeed;
     }
 
     @JsonProperty("aggregate_cpu_speed")
-    public void setAggregateCpuSpeed(Integer aggregateCpuSpeed) {
+    public void setAggregateCpuSpeed(Long aggregateCpuSpeed) {
         this.aggregateCpuSpeed = aggregateCpuSpeed;
     }
 
     @JsonProperty("aggregate_memory")
-    public Integer getAggregateMemory() {
+    public Long getAggregateMemory() {
         return aggregateMemory;
     }
 
     @JsonProperty("aggregate_memory")
-    public void setAggregateMemory(Integer aggregateMemory) {
+    public void setAggregateMemory(Long aggregateMemory) {
         this.aggregateMemory = aggregateMemory;
     }
 
     @JsonProperty("effective_cpu")
-    public Integer getEffectiveCpu() {
+    public Long getEffectiveCpu() {
         return effectiveCpu;
     }
 
     @JsonProperty("effective_cpu")
-    public void setEffectiveCpu(Integer effectiveCpu) {
+    public void setEffectiveCpu(Long effectiveCpu) {
         this.effectiveCpu = effectiveCpu;
     }
 
     @JsonProperty("effective_memory")
-    public Integer getEffectiveMemory() {
+    public Long getEffectiveMemory() {
         return effectiveMemory;
     }
 
     @JsonProperty("effective_memory")
-    public void setEffectiveMemory(Integer effectiveMemory) {
+    public void setEffectiveMemory(Long effectiveMemory) {
         this.effectiveMemory = effectiveMemory;
     }
 
     @JsonProperty("aggregate_vm_cpus")
-    public Integer getAggregateVmCpus() {
+    public Long getAggregateVmCpus() {
         return aggregateVmCpus;
     }
 
     @JsonProperty("aggregate_vm_cpus")
-    public void setAggregateVmCpus(Integer aggregateVmCpus) {
+    public void setAggregateVmCpus(Long aggregateVmCpus) {
         this.aggregateVmCpus = aggregateVmCpus;
     }
 
     @JsonProperty("aggregate_vm_memory")
-    public Integer getAggregateVmMemory() {
+    public Long getAggregateVmMemory() {
         return aggregateVmMemory;
     }
 
     @JsonProperty("aggregate_vm_memory")
-    public void setAggregateVmMemory(Integer aggregateVmMemory) {
+    public void setAggregateVmMemory(Long aggregateVmMemory) {
         this.aggregateVmMemory = aggregateVmMemory;
     }
 
@@ -186,12 +186,12 @@ public class EmsCluster {
     }
 
     @JsonProperty("drs_migration_threshold")
-    public Integer getDrsMigrationThreshold() {
+    public Long getDrsMigrationThreshold() {
         return drsMigrationThreshold;
     }
 
     @JsonProperty("drs_migration_threshold")
-    public void setDrsMigrationThreshold(Integer drsMigrationThreshold) {
+    public void setDrsMigrationThreshold(Long drsMigrationThreshold) {
         this.drsMigrationThreshold = drsMigrationThreshold;
     }
 
@@ -216,32 +216,32 @@ public class EmsCluster {
     }
 
     @JsonProperty("ha_max_failures")
-    public Integer getHaMaxFailures() {
+    public Long getHaMaxFailures() {
         return haMaxFailures;
     }
 
     @JsonProperty("ha_max_failures")
-    public void setHaMaxFailures(Integer haMaxFailures) {
+    public void setHaMaxFailures(Long haMaxFailures) {
         this.haMaxFailures = haMaxFailures;
     }
 
     @JsonProperty("total_direct_vms")
-    public Integer getTotalDirectVms() {
+    public Long getTotalDirectVms() {
         return totalDirectVms;
     }
 
     @JsonProperty("total_direct_vms")
-    public void setTotalDirectVms(Integer totalDirectVms) {
+    public void setTotalDirectVms(Long totalDirectVms) {
         this.totalDirectVms = totalDirectVms;
     }
 
     @JsonProperty("total_direct_miq_templates")
-    public Integer getTotalDirectMiqTemplates() {
+    public Long getTotalDirectMiqTemplates() {
         return totalDirectMiqTemplates;
     }
 
     @JsonProperty("total_direct_miq_templates")
-    public void setTotalDirectMiqTemplates(Integer totalDirectMiqTemplates) {
+    public void setTotalDirectMiqTemplates(Long totalDirectMiqTemplates) {
         this.totalDirectMiqTemplates = totalDirectMiqTemplates;
     }
